M42 model format changed to binary, new features
* Now supports hard and soft edges
* Texture coordinates work
* Vertices are duplicated to allow for these features
* Not yet implemented: only duplicate vertices if necessary

+++ src/utils/tgaloader.cpp
@@ -106,7 +106,7 @@ bool TgaLoader::loadUncompressed(FILE * file)
 	}
 
 	// Byte Swapping Optimized By Steve Thomas
-	for (GLuint cswap = 0; cswap < (int) tga.imageSize; cswap += tga.bytesPerPixel)
+	for (GLuint cswap = 0; cswap < tga.imageSize; cswap += tga.bytesPerPixel)
 	{
 		imageData[cswap] ^= imageData[cswap + 2] ^= imageData[cswap] ^=
 				imageData[cswap + 2];
